# Speed up pending changes lookups using ahash