--- a/lib/segment/Cargo.toml
+++ b/lib/segment/Cargo.toml
@@ -196,6 +196,10 @@ name = "in_memory_id_tracker"
 harness = false
 
 [[bench]]
+name = "mmap_bitslice_buffered_update_wrapper"
+harness = false
+
+[[bench]]
 name = "scorer_mmap"
 harness = false