--- a/lib/segment/src/common/rocksdb_buffered_delete_wrapper.rs
+++ b/lib/segment/src/common/rocksdb_buffered_delete_wrapper.rs
@@ -1,7 +1,7 @@
-use std::collections::HashSet;
 use std::mem;
 use std::sync::Arc;
 
+use ahash::AHashSet;
 use parking_lot::{Mutex, RwLock};
 use rocksdb::DB;
 
@@ -21,7 +21,7 @@ use crate::common::rocksdb_wrapper::{DatabaseColumnWrapper, LockedDatabaseColumn
 #[derive(Debug)]
 pub struct DatabaseColumnScheduledDeleteWrapper {
     db: DatabaseColumnWrapper,
-    deleted_pending_persistence: Arc<Mutex<HashSet<Vec<u8>>>>,
+    deleted_pending_persistence: Arc<Mutex<AHashSet<Vec<u8>>>>,
 }
 
 impl Clone for DatabaseColumnScheduledDeleteWrapper {
@@ -37,7 +37,7 @@ impl DatabaseColumnScheduledDeleteWrapper {
     pub fn new(db: DatabaseColumnWrapper) -> Self {
         Self {
             db,
-            deleted_pending_persistence: Arc::new(Mutex::new(HashSet::new())),
+            deleted_pending_persistence: Arc::new(Mutex::new(AHashSet::new())),
         }
     }
 
@@ -124,7 +124,7 @@ impl DatabaseColumnScheduledDeleteWrapper {
 
 pub struct LockedDatabaseColumnScheduledDeleteWrapper<'a> {
     base: LockedDatabaseColumnWrapper<'a>,
-    deleted_pending_persistence: &'a Mutex<HashSet<Vec<u8>>>,
+    deleted_pending_persistence: &'a Mutex<AHashSet<Vec<u8>>>,
 }
 
 impl LockedDatabaseColumnScheduledDeleteWrapper<'_> {
@@ -138,7 +138,7 @@ impl LockedDatabaseColumnScheduledDeleteWrapper<'_> {
 
 pub struct DatabaseColumnScheduledDeleteIterator<'a> {
     base: DatabaseColumnIterator<'a>,
-    deleted_pending_persistence: &'a Mutex<HashSet<Vec<u8>>>,
+    deleted_pending_persistence: &'a Mutex<AHashSet<Vec<u8>>>,
 }
 
 impl Iterator for DatabaseColumnScheduledDeleteIterator<'_> {