--- a/lib/segment/src/payload_storage/in_memory_payload_storage_impl.rs
+++ b/lib/segment/src/payload_storage/in_memory_payload_storage_impl.rs
@@ -1,4 +1,3 @@
-use std::collections::HashMap;
 use std::path::PathBuf;
 
 use common::counter::hardware_counter::HardwareCounterCell;
@@ -92,7 +91,7 @@ impl PayloadStorage for InMemoryPayloadStorage {
     }
 
     fn wipe(&mut self, _: &HardwareCounterCell) -> OperationResult<()> {
-        self.payload = HashMap::new();
+        self.payload = ahash::AHashMap::new();
         Ok(())
     }
 
@@ -188,7 +187,7 @@ mod tests {
                 payload.borrow().as_ref().cloned().unwrap()
             }),
             Some(&id_tracker),
-            &HashMap::new(),
+            &std::collections::HashMap::new(),
             &query,
             0,
             &IndexesMap::new(),