--- a/lib/segment/src/payload_storage/payload_storage_enum.rs
+++ b/lib/segment/src/payload_storage/payload_storage_enum.rs
@@ -153,6 +153,7 @@ impl PayloadStorage for PayloadStorageEnum {
         }
     }
 
+    #[cfg(test)]
     fn wipe(&mut self, hw_counter: &HardwareCounterCell) -> OperationResult<()> {
         match self {
             #[cfg(feature = "testing")]