A very simple change - replace the standard [`HashMap`](https://doc.rust-lang.org/std/collections/struct.HashMap.html) with [`AHashMap`](https://docs.rs/ahash/latest/ahash/struct.AHashMap.html). Or more specifically, hash using [`ahash`](https://docs.rs/ahash).

It changes the hashmap in the mmap/rocksdb update buffer types where it lists changes pending flush. Its also applied to some surrounding types/functions. There are more places we can benefit from this, but lets take it step by step.

Ahash is much faster, at least on small numeric keys like we're using in many places. Here's a generic benchmark of <ins>sip13</ins> (used by standard HashMap) versus <ins>ahash</ins>:

![image](https://github.com/user-attachments/assets/8f290057-63b1-406e-b1a2-d4c6b0f8b139)
<small>[source](https://github.com/tkaitchuck/aHash/blob/7d5c661a74b12d5bc5448b0b83fdb429190db1a3/README.md#comparison-with-other-hashers)</small>

I've added a criterion benchmark that tests the change in the buffered mmap bitslice. I expect the effect to be similar for the other types.

Here's a bench for 100 iterations on look ups with 1 million pending changes:

```bash
# Before PR
$ cargo bench --bench mmap_bitslice_buffered_update_wrapper
mmap-bitslice-buffered-update-wrapper/lookup-with-pending-changes
                        time:   [43.298 ms 43.611 ms 43.949 ms]

# After PR
$ cargo bench --bench mmap_bitslice_buffered_update_wrapper
mmap-bitslice-buffered-update-wrapper/lookup-with-pending-changes
                        time:   [30.691 ms 30.881 ms 31.113 ms]
                        change: [-29.886% -29.189% -28.441%] (p = 0.00 < 0.05)
                        Performance has improved.
```

![Graph for 100 criterion samples](https://github.com/user-attachments/assets/443f8417-bada-4c87-a0b6-dbfef9eda876)

Red shows the lookup timing distribution for the type before this PR. Blue shows the timings after this PR, being about 30% faster.

Note that in the case of the pending changes list, it may be empty a lot of the time. In that case there is no time difference. The improvement in lookup time will get more and more significant as the pending changes list grows.

### All Submissions:

* [x] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### New Feature Submissions:

1. [x] Does your submission pass tests?
2. [x] Have you formatted your code locally using `cargo +nightly fmt --all` command prior to submission?
3. [x] Have you checked your code using `cargo clippy --all --all-features` command?

### Changes to Core Features:

* [x] Have you added an explanation of what your changes do and why you'd like us to include them?
* [x] Have you written new tests for your core changes, as applicable?
* [x] Have you successfully ran tests with your changes locally?